Revert "LPS-168476 Extract PreviewButton"
This reverts commit c8ef1ad3d8d263e748bfe4c0c3b403f20858d912.
